+++ app/controllers/staff/activity_children_controller.rb
@@ -23,8 +23,6 @@ class Staff::ActivityChildrenController < Staff::BaseController
 
     activity.save!
 
-    activity.create_activity key: "activity.create", owner: current_user
-
     redirect_to activity_step_path(activity.id, activity.form_state)
   end
 end