+++ app/controllers/staff/organisations_controller.rb
@@ -70,7 +70,6 @@ class Staff::OrganisationsController < Staff::BaseController
 
     if @organisation.valid?
       @organisation.save
-      @organisation.create_activity key: "organisation.create", owner: current_user
       flash[:notice] = t("action.organisation.create.success")
       redirect_to organisation_path(@organisation)
     else
@@ -91,7 +90,6 @@ class Staff::OrganisationsController < Staff::BaseController
 
     if @organisation.valid?
       @organisation.save
-      @organisation.create_activity key: "organisation.update", owner: current_user
       flash[:notice] = t("action.organisation.update.success")
       redirect_to organisation_path(@organisation)
     else