+++ app/controllers/staff/programmes_controller.rb
@@ -5,8 +5,6 @@ class Staff::ProgrammesController < Staff::ActivitiesController
     @programme = CreateProgrammeActivity.new(organisation_id: organisation_id, fund_id: fund_id).call
     authorize @programme
 
-    @programme.create_activity key: "activity.create", owner: current_user
-
     redirect_to activity_step_path(@programme.id, @programme.form_state)
   end
 end