+++ app/models/external_income.rb
@@ -1,5 +1,4 @@
 class ExternalIncome < ApplicationRecord
-  include PublicActivity::Common
   include HasFinancialQuarter
 
   belongs_to :organisation