+++ spec/features/staff/users_can_edit_a_budget_spec.rb
@@ -3,16 +3,17 @@ RSpec.describe "Users can edit a budget" do
 
   context "when signed in beis user" do
     let(:user) { create(:beis_user) }
+    let!(:activity) { create(:programme_activity, organisation: user.organisation) }
+    let!(:budget) { create(:budget, parent_activity: activity, value: "10") }
 
-    scenario "a budget can be successfully edited" do
-      activity = create(:programme_activity, organisation: user.organisation)
-      budget = create(:budget, parent_activity: activity, value: "10")
-
+    before do
       visit organisation_activity_path(user.organisation, activity)
       within("##{budget.id}") do
         click_on t("default.link.edit")
       end
+    end
 
+    scenario "a budget can be successfully edited" do
       fill_in "budget[value]", with: "20"
       click_on t("default.button.submit")
 
@@ -21,22 +22,33 @@ RSpec.describe "Users can edit a budget" do
       within("##{budget.id}") do
         expect(page).to have_content("20.00")
       end
+    end
+
+    scenario "a budget can be successfully deleted" do
+      click_on t("default.button.delete")
+
+      expect(page).to have_content(t("action.budget.destroy.success"))
+      expect(page).to_not have_content("10.00")
+
+      expect { budget.reload }.to raise_error { ActiveRecord::RecordNotFound }
     end
   end
 
   context "when signed in as delivery partner user" do
     let(:user) { create(:delivery_partner_user) }
+    let(:activity) { create(:project_activity, organisation: user.organisation) }
+    let(:report) { create(:report, state: :active, organisation: user.organisation, fund: activity.associated_fund) }
 
-    scenario "a budget can be successfully edited" do
-      activity = create(:project_activity, organisation: user.organisation)
-      report = create(:report, state: :active, organisation: user.organisation, fund: activity.associated_fund)
-      budget = create(:budget, parent_activity: activity, value: "10", report: report)
+    let!(:budget) { create(:budget, parent_activity: activity, value: "10", report: report) }
 
+    before do
       visit organisation_activity_path(user.organisation, activity)
       within("##{budget.id}") do
         click_on t("default.link.edit")
       end
+    end
 
+    scenario "a budget can be successfully edited" do
       fill_in "budget[value]", with: "20"
       click_on t("default.button.submit")
 
@@ -46,37 +58,16 @@ RSpec.describe "Users can edit a budget" do
       end
     end
 
-    scenario "budget update is tracked with public_activity" do
-      activity = create(:project_activity, organisation: user.organisation)
-      report = create(:report, state: :active, organisation: user.organisation, fund: activity.associated_fund)
-      budget = create(:budget, parent_activity: activity, budget_type: "direct", value: "10", report: report)
+    scenario "a budget can be successfully deleted" do
+      click_on t("default.button.delete")
 
-      PublicActivity.with_tracking do
-        visit organisation_activity_path(user.organisation, activity)
-        within("##{budget.id}") do
-          click_on t("default.link.edit")
-        end
+      expect(page).to have_content(t("action.budget.destroy.success"))
+      expect(page).to_not have_content("10.00")
 
-        fill_in "budget[value]", with: "20"
-        click_on t("default.button.submit")
-
-        budget = Budget.last
-        auditable_event = PublicActivity::Activity.find_by(trackable_id: budget.id)
-        expect(auditable_event.key).to eq "budget.update"
-        expect(auditable_event.owner_id).to eq user.id
-      end
+      expect { budget.reload }.to raise_error { ActiveRecord::RecordNotFound }
     end
 
     scenario "validation errors work as expected" do
-      activity = create(:project_activity, organisation: user.organisation)
-      report = create(:report, state: :active, organisation: user.organisation, fund: activity.associated_fund)
-      budget = create(:budget, parent_activity: activity, value: "10", report: report)
-
-      visit organisation_activity_path(user.organisation, activity)
-      within("##{budget.id}") do
-        click_on t("default.link.edit")
-      end
-
       fill_in "budget[value]", with: ""
 
       click_on t("default.button.submit")